requirements.txt
# Core
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0

# Twilio for calls
twilio==8.10.0

# OpenAI for Whisper and ChatGPT
openai==1.12.0

# ElevenLabs for TTS
elevenlabs==0.2.24

# Audio processing
pydub==0.25.1